Add Popover component
